requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2

psycopg2-binary==2.9.9
python-dotenv==1.0.0

gunicorn==21.2.0

itsdangerous==2.1.2
click==8.1.7

# Optional utilities (safe to keep)
requests==2.31.0
pillow==10.2.0
